use a latest release as a reference
