+++ .github/workflows/test.yml
@@ -31,6 +31,5 @@ jobs:
           FORCE_COLOR: 1
           ENABLE_GAS_REPORT: true
       - run: npm run test:inheritance
-      - run: npm run layout:test
       - name: Print gas report
         run: cat gas-report.txt
