+++ .github/workflows/transpile.yml
@@ -33,4 +33,5 @@ jobs:
         if: steps.cache.outputs.cache-hit != 'true'
       - run: bash scripts/upgradeable/git-user-config.sh
       - run: bash scripts/upgradeable/transpile-onto.sh ${{ steps.branch.outputs.name }} origin/${{ steps.branch.outputs.name }}
+      - run: bash scripts/upgradeable/layout-test.sh
       - run: git push origin ${{ steps.branch.outputs.name }}
